+++ src/panoptes/utils/rs232.py
@@ -3,12 +3,14 @@ import time
 from contextlib import suppress
 
 import serial
+from deprecated import deprecated
 from loguru import logger
 from panoptes.utils import error
 from panoptes.utils import serializers
 from serial.tools.list_ports import comports as get_comports
 
 
+@deprecated(reason='Use panoptes.utils.serial.device')
 def get_serial_port_info():
     """Returns the serial ports defined on the system.
 
@@ -18,6 +20,7 @@ def get_serial_port_info():
     return sorted(get_comports(), key=operator.attrgetter('device'))
 
 
+@deprecated(reason='Use panoptes.utils.serial.device')
 def find_serial_port(vendor_id, product_id, return_all=False):  # pragma: no cover
     """Finds the serial port that matches the given vendor and product id.
 
@@ -56,6 +59,7 @@ def find_serial_port(vendor_id, product_id, return_all=False):  # pragma: no cov
             f'No serial ports for vendor_id={vendor_id:x} and product_id={product_id:x}')
 
 
+@deprecated(reason='Use panoptes.utils.serial.device')
 class SerialData(object):
     """SerialData wraps a PySerial instance for reading from and writing to a serial device.
 
@@ -67,13 +71,10 @@ class SerialData(object):
 
     .. doctest::
 
-        >>> import serial
+        # Register our serial simulators by importing protocol.
+        >>> from panoptes.utils.serial.handlers import protocol_buffers
 
-        # Register our serial simulators
-        >>> serial.protocol_handler_packages.append('panoptes.utils.serial_handlers')
-        >>> from panoptes.utils.serial_handlers import protocol_buffers as pb
-
-        # Import our serial utils
+        # Import our serial utils.
         >>> from panoptes.utils.rs232 import SerialData
 
         # Connect to our fake buffered device
@@ -81,7 +82,7 @@ class SerialData(object):
 
         # Note: A manual reset is currently required because implementation is not complete.
         # See https://github.com/panoptes/POCS/issues/758 for details.
-        >>> pb.ResetBuffers()
+        >>> protocol_buffers.reset_serial_buffers()
         >>> device_listener.is_connected
         True
 
@@ -89,7 +90,7 @@ class SerialData(object):
         'buffers://'
 
         # Device sends event
-        >>> pb.SetRBufferValue(b'emit event')
+        >>> protocol_buffers.set_serial_read_buffer(b'emit event')
 
         # Listen for event
         >>> device_listener.read()
@@ -97,11 +98,12 @@ class SerialData(object):
 
         >>> device_listener.write('ack event')
         9
-        >>> pb.GetWBufferValue()
+        >>> protocol_buffers.get_serial_write_buffer()
         b'ack event'
 
         # Remove custom handlers
-        >>> serial.protocol_handler_packages.remove('panoptes.utils.serial_handlers')
+        >>> import serial
+        >>> serial.protocol_handler_packages.remove('panoptes.utils.serial.handlers')
     """
 
     def __init__(self,