+++ tests/test_rs232.py
@@ -1,14 +1,12 @@
 import io
+
 import pytest
 import serial
-from serial import serialutil
-
 from panoptes.utils import error
 from panoptes.utils import rs232
-
-from panoptes.utils.serial_handlers import NoOpSerial
-from panoptes.utils.serial_handlers import protocol_buffers
-from panoptes.utils.serial_handlers import protocol_hooked
+from panoptes.utils.serial.handlers import protocol_buffers, protocol_hooked
+from panoptes.utils.serial.handlers.protocol_no_op import NoOpSerial
+from serial.serialutil import PortNotOpenError
 
 
 def test_port_discovery():
@@ -33,23 +31,13 @@ def test_non_existent_device():
     assert not ser.is_connected
 
 
-def test_detect_uninstalled_scheme():
-    """If our handlers aren't installed, will detect unknown scheme."""
-    # See https://pythonhosted.org/pyserial/url_handlers.html#urls for info on the
-    # standard schemes that are supported by PySerial.
-    with pytest.raises(ValueError):
-        # The no_op scheme references one of our test handlers, but it shouldn't be
-        # accessible unless we've added our package to the list to be searched.
-        rs232.SerialData(port='no_op://')
-
-
 @pytest.fixture(scope='function')
 def handler():
     # Install our package that contain the test handlers.
-    serial.protocol_handler_packages.append('panoptes.utils.serial_handlers')
+    serial.protocol_handler_packages.append('panoptes.utils.serial.handlers')
     yield True
     # Remove that package.
-    serial.protocol_handler_packages.remove('panoptes.utils.serial_handlers')
+    serial.protocol_handler_packages.remove('panoptes.utils.serial.handlers')
 
 
 def test_detect_bogus_scheme(handler):
@@ -107,9 +95,8 @@ def test_basic_no_op(handler):
 
 
 def test_basic_io(handler):
-    protocol_buffers.ResetBuffers(b'abc\r\ndef\n')
-    ser = rs232.SerialData(port='buffers://', open_delay=0.01, retry_delay=0.01,
-                           retry_limit=2)
+    protocol_buffers.reset_serial_buffers(b'abc\r\ndef\n')
+    ser = rs232.SerialData(port='buffers://', open_delay=0.01, retry_delay=0.01, retry_limit=2)
 
     # Peek inside, it should have a BuffersSerial instance as member ser.
     assert isinstance(ser.ser, protocol_buffers.BuffersSerial)
@@ -128,10 +115,10 @@ def test_basic_io(handler):
     assert 5 == ser.write('def\r\n')
     assert 6 == ser.write('done\r\n')
 
-    assert b'def\r\ndone\r\n' == protocol_buffers.GetWBufferValue()
+    assert b'def\r\ndone\r\n' == protocol_buffers.get_serial_write_buffer()
 
     # If we add more to the read buffer, we can read again.
-    protocol_buffers.SetRBufferValue(b'line1\r\nline2\r\ndangle')
+    protocol_buffers.set_serial_read_buffer(b'line1\r\nline2\r\ndangle')
     assert 'line1\r\n' == ser.read(retry_delay=10, retry_limit=20)
     assert 'line2\r\n' == ser.read(retry_delay=10, retry_limit=20)
     assert 'dangle' == ser.read(retry_delay=10, retry_limit=20)
@@ -145,14 +132,13 @@ class HookedSerialHandler(NoOpSerial):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.r_buffer = io.BytesIO(
-            b"{'a': 12, 'b': [1, 2, 3, 4], 'c': {'d': 'message'}}\r\n")
+        self.r_buffer = io.BytesIO(b"{'a': 12, 'b': [1, 2, 3, 4], 'c': {'d': 'message'}}\r\n")
 
     @property
     def in_waiting(self):
         """The number of input bytes available to read immediately."""
         if not self.is_open:
-            raise serialutil.portNotOpenError
+            raise PortNotOpenError
         total = len(self.r_buffer.getbuffer())
         avail = total - self.r_buffer.tell()
         # If at end of the stream, reset the stream.
@@ -162,11 +148,7 @@ class HookedSerialHandler(NoOpSerial):
         return avail
 
     def open(self):
-        """Open port.
-
-        Raises:
-            SerialException if the port cannot be opened.
-        """
+        """Open port."""
         self.is_open = True
 
     def close(self):
@@ -176,14 +158,14 @@ class HookedSerialHandler(NoOpSerial):
     def read(self, size=1):
         """Read until the end of self.r_buffer, then seek to beginning of self.r_buffer."""
         if not self.is_open:
-            raise serialutil.portNotOpenError
+            raise PortNotOpenError
         # If at end of the stream, reset the stream.
         return self.r_buffer.read(min(size, self.in_waiting))
 
     def write(self, data):
-        """Write data to bitbucket."""
+        """Write noop."""
         if not self.is_open:
-            raise serialutil.portNotOpenError
+            raise PortNotOpenError
         return len(data)
 
 
@@ -212,8 +194,7 @@ def test_hooked_io(handler):
         assert reading[1] == line
 
     # Can write to the "device" many times.
-    line = 'abcdefghijklmnop' * 30
-    line = line + '\r\n'
+    line = f'{"foobar" * 30}\r\n'
     for n in range(20):
         assert len(line) == ser.write(line)